fix(ios): thread chunk_a -> chunk_b hidden as fp32 to dodge iPhone CPU kernel bug
iPhone 17 Pro benchmark revealed a new iOS 26.1 Core ML CPU kernel bug:
feeding the fp16 'hidden' MLMultiArray output of chunk_a straight into
chunk_b's fp16 'hidden_in' produces wildly wrong values on iPhone CPU
only. Measured on-device with the fp16 handoff:
  out[0..5]=5.84, 3.41, +3.65, -0.18, -2.26
  ref[0..5]=5.93, 3.87, -0.68, +1.03, +0.38
The 3rd value flips sign (-0.68 -> +3.65) and argmax becomes a random
~100k token id (oracle is 3629). Mac CPU handled the same fp16 pair
correctly (cos=0.9999). So this is specific to iOS CPU (tested 26.1),
similar in flavor to the earlier layer_norm NaN and BNNS IR overruns —
the iPhone CPU path has accumulated bugs that Mac CPU doesn't share.

Fix: declare chunk_a output 'hidden' and chunk_b input 'hidden_in' as
fp32 instead of fp16. +128 KB (64 * 1024 * 2) per call, negligible
at iPhone memory scales. Mac CPU parity stays at cos=0.9999 mean.
ANE / All compute units should keep working the same — they handle
fp32 hidden fine; coremltools inserts an implicit cast if the ANE
kernel needs fp16 internally.

iPhone retest expected:
  CPU only chunked prefill: 0% top-1 -> 100% top-1 (recovery)
  CPU+ANE  chunked prefill: still 80% top-1 (ANE fp16 ceiling, unchanged)
  All       chunked prefill: still 80% top-1

If this validates, the CPU-only shipping path is restored on iPhone
with chunked prefill, which is the baseline before ANE optimisations
can be re-layered.

+++ conversion/build_qwen35_chunks.py
@@ -200,10 +200,15 @@ def main():
     out_dir = Path(args.out_dir)
     out_dir.mkdir(parents=True, exist_ok=True)
 
+    # Hidden tensor between chunks as fp32 to dodge an iPhone Core ML CPU
+    # kernel bug: feeding an fp16 MLMultiArray output from one model into
+    # another model's fp16 input produces wrong values on iPhone CPU only
+    # (tested on iOS 26.1, Mac CPU handled the same pair correctly).
+    # fp32 hidden = +128 KB at seq=64/hidden=1024, negligible.
     convert_chunk(
         chunk_a,
         input_specs=[{"name": "input_ids", "shape": (1, args.seq_len), "dtype": np.int32}],
-        output_specs=[{"name": "hidden", "dtype": np.float16}],
+        output_specs=[{"name": "hidden", "dtype": np.float32}],
         out_path=out_dir / "qwen3_5_chunk_a.mlpackage",
         label=f"chunk_a (embed + layers [0, {args.split}))",
     )
@@ -211,7 +216,7 @@ def main():
         chunk_b,
         input_specs=[{"name": "hidden_in",
                       "shape": (1, args.seq_len, cfg.hidden_size),
-                      "dtype": np.float16}],
+                      "dtype": np.float32}],
         output_specs=[{"name": "logits", "dtype": np.float32}],
         out_path=out_dir / "qwen3_5_chunk_b.mlpackage",
         label=f"chunk_b (layers [{args.split}, {cfg.num_hidden_layers}) + lm_head)",
